Added fashionHouseApi and DesignerApi unit tests

+++ ClothesShop/Controllers/ClothesApiController.cs
@@ -5,13 +5,13 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace ClothesShop.Controllers
 {
-    [Route("api/designer")]
+    [Route("api/clothes")]
     [ApiController]
-    public class DesignersApiController : ControllerBase
+    public class ClothesApiController : ControllerBase
     {
-        private readonly IDesignerService _service;
+        private readonly IClothesService _service;
 
-        public DesignersApiController(IDesignerService service)
+        public ClothesApiController(IClothesService service)
         {
             _service = service;
         }
@@ -28,7 +28,7 @@ namespace ClothesShop.Controllers
         public async Task<IActionResult> Get(int id)
         {
             var result = await _service.GetByIdAsync(id);
-            if(result == null)
+            if (result == null)
             {
                 return NotFound();
             }
@@ -39,38 +39,38 @@ namespace ClothesShop.Controllers
         [Route("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
-
-            var isDeleted = await _service.DeleteAsync(id);
-            if (isDeleted)
+            var result = await _service.GetByIdAsync(id);
+            if (result != null)
             {
+                await _service.DeleteAsync(id);
                 return Ok();
             }
             return NotFound();
         }
 
         [HttpPost]
-        public async Task<IActionResult> Create([FromBody] Designer designer)
+        public async Task<IActionResult> Create([FromBody] NewClothesVm clothes)
         {
-            if(!ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
                 return BadRequest();
             }
-            await _service.AddAsync(designer);
-            return Created($"/api/designer/{designer.Id}", designer);
+            await _service.AddNewClothes(clothes);
+            return Created($"/api/book/{clothes.Id}", clothes);
         }
 
         [HttpPut]
         [Route("{id}")]
-        public async Task<IActionResult> Update([FromRoute]int id,[FromBody]Designer designer)
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] NewClothesVm clothes)
         {
-            var foundDesigner = await _service.GetByIdAsync(id);
-            if (foundDesigner == null)
+            var found = await _service.GetByIdAsync(id);
+            if (found == null)
             {
                 return NotFound();
             }
             if (ModelState.IsValid)
             {
-                await _service.UpdateAsync(id, designer);
+                await _service.UpdateClothesAsync(clothes);
                 return Ok();
             }
             return BadRequest();
